% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyRun.R
\name{EasyRun}
\alias{EasyRun}
\title{Easily complete the recommendation of imputation methods}
\usage{
EasyRun(
  RawData,
  GroupAsize,
  GroupBsize,
  DEratio = 0.2,
  logFC = 1,
  Pair = TRUE,
  ExtraMethod = NULL
)
}
\arguments{
\item{RawData}{A two-dimensional matrix with missing values}

\item{GroupAsize}{Sample size of group A}

\item{GroupBsize}{Sample size of group B}

\item{logFC}{Fold Change after log2 conversion for simulation}

\item{Pair}{Paired experimental design or not}

\item{ExtraMethod}{A vector for extra methods (You are allowed to use imputation functions not included in this package, just enter their names). If there is no extra method, please enter NULL}

\item{DERatio}{Differential expression ratio in simulation matrix}
}
\value{
A list containing the following components:
\item{Total}{    Overall result}
\item{NRMSE}{    Top 3 with NRMSE}
\item{BLCI}{    Top 3 with BLCI}
}
\description{
An easy-to-use way to recommend optimal imputation method
}
\note{
Require \code{\link{imputeLCMD}}
}
\examples{
##################################### a simulated example #################
rm(list = ls())
library('imputeLCMD')
set.seed(2020)
a <- matrix(data = rnorm(100000), ncol = 100) # Generate random matrix
a[sample(length(a), 10000)] <- NA # Generate missing values randomly
rownames(a) <- seq(1, 1000, 1) # Matrix must have row names
res <- EasyRun(RawData = a, GroupAsize = 50, GroupBsize = 50, DEratio = 0.2, logFC = 1, Pair = T, ExtraMethod = NULL) # Evaluate input Imputation methods

}
