% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputePower.R
\name{ImputePower}
\alias{ImputePower}
\title{Estimate the power of imputation methods based on input data}
\usage{
ImputePower(
  RawData,
  GroupAsize,
  GroupBsize,
  DERatio = 0.2,
  logFC = 1,
  ReTimes = NULL,
  Pair,
  MNARratio,
  ImputeMethod = c("kNN_fill", "SVD_fill", "MLE_fill", "Mean_fill", "QRILC_fill",
    "LLS_fill", "BPCA_fill", "Min_fill")
)
}
\arguments{
\item{RawData}{A two-dimensional matrix with missing values}

\item{GroupAsize}{Sample size of group A}

\item{GroupBsize}{Sample size of group B}

\item{DERatio}{Differential expression ratio in simulation matrix}

\item{logFC}{Fold Change after log2 conversion for simulation}

\item{ReTimes}{Repeat times. To eliminate random error. The smaller the more accurate but more time consuming. You can enter NULL and we will automatically calculate the repeat times based on the matrix size}

\item{Pair}{Paired experimental design or not}

\item{MNARratio}{MNAR ratio of the input matrix}

\item{ImputeMethod}{Name of imputation methods for evaluation and comparison (You can use imputation functions not included in this package,just enter the name)}
}
\value{


A list containing the following components:
\item{Total}{    Overall result}
\item{NRMSE}{    NRMSE of each imputation method}
\item{BLCI}{    BLCI of each imputation method}
}
\description{
By establishing simulation matrices, the effectiveness of each Imputation method is evaluated from the two levels of NRMSE and BLCI (TPR + TNR - 1). First generate the corresponding simulation matrix based on the input matrix and dataset features. Then after generating the missing values, use each Imputation method to be evaluated to impute the missing values. Compare the initial simulation matrix with the filled simulation matrix to get the result. Repeat to eliminate random errors
}
\note{
Require \code{\link{imputeLCMD}}
}
\examples{
##################################### a simulated example ###################
rm(list = ls())
library('imputeLCMD')
set.seed(2020)
a <- matrix(data = rnorm(100000), ncol = 100) # Generate random matrix
a[sample(length(a), 10000)] <- NA # Generate missing values randomly
rownames(a) <- seq(1, 1000, 1) # Matrix must have row names
res <- ImputePower(RawData = a, GroupAsize = 50, GroupBsize = 50, DERatio = 0.2, logFC = 1, ReTimes = 30, Pair = T, MNARratio = 0,
                   ImputeMethod = c('kNN_fill', 'SVD_fill', 'MLE_fill', 'Mean_fill', 'QRILC_fill', 'LLS_fill', 'BPCA_fill', 'Min_fill')) # Evaluate input Imputation methods



}
