% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN_fill.R
\name{kNN_fill}
\alias{kNN_fill}
\title{Wrapper function for KNN imputation}
\usage{
kNN_fill(RawData, K = 10)
}
\arguments{
\item{RawData}{Matrix with missing values}

\item{K}{The number of neighbors used to infer the missing data.}
}
\value{
A complete matrix with missing values imputed by kNN
}
\description{
This function is a wrapper for the \code{\link{impute.knn}} function, which performs KNN imputation on a data matrix containing missing values. Impute missing values by weighted average abundance of k most similar peptides
}
\note{
Require \code{\link{imputeLCMD}}
}
\examples{
##################################### a simulated example #################
rm(list = ls())
set.seed(2020)
a <- matrix(data = rnorm(100000), ncol = 100) # Generate random matrix
a[sample(length(a), 10000)] <- NA # Generate missing values randomly
if(require(imputeLCMD)){
  a_1 <- kNN_fill(a)
}

}
\seealso{
\code{\link{BPCA_fill}}, \code{\link{LLS_fill}},  \code{\link{Mean_fill}}, \code{\link{QRILC_fill}}, \code{\link{SVD_fill}}, \code{\link{MLE_fill}}, \code{\link{Min_fill}}
}
