% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_fill.R
\name{Mean_fill}
\alias{Mean_fill}
\title{Wrapper function for imputation using mean of each row}
\usage{
Mean_fill(RawData)
}
\arguments{
\item{RawData}{Matrix with missing values}
}
\value{
A complete matrix with missing values imputed by mean of each row
}
\description{
This function performs mean imputation on a data matrix containing missing values. Use the average of the non-missing elements in each row.
}
\examples{
##################################### a simulated example #################
rm(list = ls())
set.seed(2020)
a <- matrix(data = rnorm(100000), ncol = 100) # Generate random matrix
a[sample(length(a), 10000)] <- NA # Generate missing values randomly
a_1 <- Mean_fill(a)

}
\seealso{
\code{\link{kNN_fill}}, \code{\link{BPCA_fill}},  \code{\link{LLS_fill}}, \code{\link{QRILC_fill}}, \code{\link{SVD_fill}}, \code{\link{MLE_fill}}, \code{\link{Min_fill}}
}
