% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissFilter.R
\name{MissFilter}
\alias{MissFilter}
\title{Function for missing filtering}
\usage{
MissFilter(RawData, MaxMiss = 0.5, byrow = TRUE)
}
\arguments{
\item{RawData}{A matrix that requires missing filtering}

\item{MaxMiss}{Maximum percentage of missing values allowed per row/columns}

\item{byrow}{Whether to filter by rows. If not, proceed by column}
}
\value{
Matrix after missing filtering
}
\description{
Filter out rows/columns where the number of missing values is more than the input value
}
\examples{
##################################### a simulated example #################
rm(list = ls())
set.seed(2020)
a <- matrix(data = rnorm(100000), ncol = 100) # Generate random matrix
a[sample(length(a), 10000)] <- NA # Generate missing values randomly
a_1 <- MissFilter(RawData = a, MaxMiss = 0.1, byrow = T) # Filter out rows with more than 50\% missing values

}
