% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNARCalc.R
\name{MNARCalc}
\alias{MNARCalc}
\title{Calculate MNAR ratio of input data}
\usage{
MNARCalc(RawData, StepSize = NULL)
}
\arguments{
\item{RawData}{A two-dimensional matrix with missing values}

\item{StepSize}{The percentage of growth for each step (from 0\% to 100\% ). The smaller the more accurate but more time consuming. You can enter NULL and we will automatically calculate the step size based on the matrix size}
}
\value{
A list containing the following components:
\item{MNARratio}{ MNAR ratio of the input matrix}
\item{D.res}{ D.value corresponding to each MNAR}
}
\description{
The MNAR ratio of the input matrix is estimated by establishing simulation matrices. Generate corresponding simulation matrices based on the missing ratio of the input data and each MNAR ratio in the given range. Then calculate the maximum vertical difference between the distributions (D.value) of each simulation matrix and the input matrix. The MNAR ratio with the smallest D.value is the estimation result.
}
\examples{
##################################### a simulated example #################
rm(list = ls())
set.seed(2020)
a <- matrix(data = rnorm(100000), ncol = 100) # Generate random matrix
a[sample(length(a), 10000)] <- NA # Generate missing values randomly
res <- MNARCalc(RawData = a, StepSize = NULL) # Calculating MNAR ratio


}
