% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVD_fill.R
\name{SVD_fill}
\alias{SVD_fill}
\title{Wrapper function for SVD-based imputation}
\usage{
SVD_fill(RawData, K = 2)
}
\arguments{
\item{RawData}{Matrix with missing values}

\item{K}{The number of PCs used.}
}
\value{
A complete matrix with missing values imputed by SVD
}
\description{
This function is a wrapper for the \code{\link{impute.wrapper.SVD}} function, which performs SVD imputation on a data matrix containing missing values. Use the k ranks of the matrix to continuously iterate to convergence.
}
\note{
Require \code{\link{imputeLCMD}}
}
\examples{
##################################### a simulated example #################
rm(list = ls())
set.seed(2020)
a <- matrix(data = rnorm(100000), ncol = 100) # Generate random matrix
a[sample(length(a), 10000)] <- NA # Generate missing values randomly
if(require(imputeLCMD)){
  a_1 <- SVD_fill(a)
}

}
\seealso{
\code{\link{BPCA_fill}}, \code{\link{LLS_fill}},  \code{\link{Mean_fill}}, \code{\link{QRILC_fill}}, \code{\link{kNN_fill}}, \code{\link{MLE_fill}}, \code{\link{Min_fill}}
}
